orchestrator-ai/requirements.txt
google-adk[a2a,extensions]>=1.26.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-qdrant>=0.2.0
qdrant-client>=1.10.0
redis>=5.0.0
python-dotenv>=1.0.0
uvicorn>=0.29.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
